+++ app/lib/data/daos/dashboard_dao.g.dart
@@ -5,6 +5,7 @@ part of 'dashboard_dao.dart';
 // ignore_for_file: type=lint
 mixin _$DashboardDaoMixin on DatabaseAccessor<GolfyDatabase> {
   $CoursesTable get courses => attachedDatabase.courses;
+  $EventsTable get events => attachedDatabase.events;
   $RoundsTable get rounds => attachedDatabase.rounds;
   $HoleResultsTable get holeResults => attachedDatabase.holeResults;
   DashboardDaoManager get managers => DashboardDaoManager(this);
@@ -15,6 +16,8 @@ class DashboardDaoManager {
   DashboardDaoManager(this._db);
   $$CoursesTableTableManager get courses =>
       $$CoursesTableTableManager(_db.attachedDatabase, _db.courses);
+  $$EventsTableTableManager get events =>
+      $$EventsTableTableManager(_db.attachedDatabase, _db.events);
   $$RoundsTableTableManager get rounds =>
       $$RoundsTableTableManager(_db.attachedDatabase, _db.rounds);
   $$HoleResultsTableTableManager get holeResults =>